Fix Android multiple spinner issue
