+++ android/src/main/java/com/react/rnspinkit/RNSpinkit.java
@@ -1,32 +1,18 @@
 package com.react.rnspinkit;
 
 
-import android.content.res.TypedArray;
 import android.graphics.Color;
-import android.graphics.Rect;
-import android.graphics.drawable.Drawable;
-import android.graphics.drawable.RotateDrawable;
 import android.support.annotation.Nullable;
 import android.util.Log;
-import android.view.ContextThemeWrapper;
-import android.view.Gravity;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.FrameLayout;
-import android.widget.LinearLayout;
-import android.widget.ListView;
-import android.widget.ProgressBar;
 import android.widget.RelativeLayout;
 
-import com.facebook.drawee.backends.pipeline.Fresco;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.uimanager.SimpleViewManager;
-import com.facebook.react.*;
 import com.facebook.react.uimanager.ThemedReactContext;
-import com.facebook.react.uimanager.ViewManager;
-import com.github.ybq.android.spinkit.SpinKitView;
 import com.facebook.react.uimanager.annotations.ReactProp;
-import com.facebook.react.views.view.ReactViewGroup;
+import com.github.ybq.android.spinkit.SpinKitView;
 import com.github.ybq.android.spinkit.sprite.Sprite;
 import com.github.ybq.android.spinkit.style.ChasingDots;
 import com.github.ybq.android.spinkit.style.Circle;
@@ -40,13 +26,17 @@ import com.github.ybq.android.spinkit.style.ThreeBounce;
 import com.github.ybq.android.spinkit.style.WanderingCubes;
 import com.github.ybq.android.spinkit.style.Wave;
 
+import java.util.HashMap;
+import java.util.Map;
+import java.util.UUID;
+
 /**
  * Created by suzuri04x2 on 2016/5/10.
  */
-public class RNSpinkit extends SimpleViewManager<SpinKitView> {
+public class RNSpinkit extends SimpleViewManager<RNSpinkitView> {
 
     ReactApplicationContext mContext;
-    Sprite mSprite = getSprite("");
+
     double mSize = 48;
 
     public RNSpinkit(ReactApplicationContext reactContext) {
@@ -59,12 +49,12 @@ public class RNSpinkit extends SimpleViewManager<SpinKitView> {
     }
 
     @Override
-    protected SpinKitView createViewInstance(ThemedReactContext reactContext) {
-        return new SpinKitView(reactContext);
+    protected RNSpinkitView createViewInstance(ThemedReactContext reactContext) {
+        return new RNSpinkitView(reactContext);
     }
 
     @ReactProp(name = "isVisible")
-    public void setIsVisible(SpinKitView view, @Nullable Boolean visible) {
+    public void setIsVisible(RNSpinkitView view, @Nullable Boolean visible) {
         if(visible)
             view.setVisibility(View.VISIBLE);
         else
@@ -72,66 +62,19 @@ public class RNSpinkit extends SimpleViewManager<SpinKitView> {
     }
 
     @ReactProp(name = "color")
-    public void setColor(SpinKitView view, @Nullable String color) {
-        try {
-            mSprite.setColor(Color.parseColor(color));
-        view.setIndeterminateDrawable(mSprite);
-        } catch(Exception err) {
-            Log.e("RNSpinkit-Err", err.toString() + "when set prop color to " + color);
-        }
+    public void setColor(RNSpinkitView view, @Nullable String color) {
+        view.setSpriteColor(color);
     }
 
     @ReactProp(name = "size")
-    public void setSize(SpinKitView view, @Nullable double size) {
-        mSize = size;
-        RelativeLayout.LayoutParams params = new RelativeLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT);
-        view.setLayoutParams(params);
+    public void setSize(RNSpinkitView view, @Nullable double size) {
+        view.setSpriteSize(size);
     }
 
     @ReactProp(name = "type")
-    public void setType(SpinKitView view, @Nullable String spinnerType) {
-        Sprite sprite = getSprite(spinnerType);
-        sprite.setColor(mSprite.getColor());
-        mSprite = sprite;
-        RelativeLayout.LayoutParams params = new RelativeLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT);
-        view.setLayoutParams(params);
-        view.setIndeterminateDrawable(sprite);
-    }
+    public void setType(RNSpinkitView view, @Nullable String spinnerType) {
+        view.setSpriteType(spinnerType);
 
-    private Sprite getSprite(String spinnerType) {
-        switch (spinnerType) {
-            case "Bounce" :
-                return new DoubleBounce();
-            case "Wave" :
-                return new Wave();
-            case "RotatingPlane" :
-                return new RotatingPlane();
-            case "WanderingCubes":
-                return new WanderingCubes();
-            case "9CubeGrid":
-                return new CubeGrid();
-            case "FadingCircleAlt" :
-                return new FadingCircle();
-            case "Pulse" :
-                return new Pulse();
-            case "ChasingDots":
-                // Add scale factor to prevent clipping
-                Sprite d = new ChasingDots();
-                d.setScale(0.85f);
-                return d;
-            case "ThreeBounce":
-                return new ThreeBounce();
-            case "Circle":
-                return new Circle();
-            case "FoldingCube":
-                // Add scale factor to prevent clipping
-                Sprite sprite = new FoldingCube();
-                sprite.setScale(0.70f);
-                return sprite;
-            default :
-                break;
-        }
-        return new RotatingPlane();
     }
 
 }
